Clean up logging behavior
Many photometry bugs/edge-cases fixed, major refactors done.
Clean up logging behavior
