+++ setup.cfg
@@ -43,9 +43,10 @@ ignore =
     # (flake8-tabs) unexpected number of tabs and spaces at start of XXX
 
 [tool:pytest]
-addopts = --strict-markers --durations=10
+addopts = --strict-markers --durations=10 -s
 testpaths = tests
 xfail_strict = True
+log_cli = True
 
 [coverage:run]
 source = ./
